[PD-7175] Single filter service (#134)
## What does this PR do?
Introduces a search service for the endpoint /filter/\<filter>. Used to fetch data for a single filter.

## Usage
This service requires the display value of a filter from the config be passed to the filter parameter. The display value was chosen as it is the only value associated with filters that is unique and present on all filters (although uniqueness is not enforced). Otherwise, it is used the same as all other services. Note this endpoint supports passing offset and num_items as part of the input, but does not provide full pagination support.

```js
import { filterSearchService } from "microsite-js/src/services/search"

//Where filter is "Job Title" and no offset/num_items are passed
filterSearchService(input, siteConfig, filter)

//Response
{
    "meta": {
        "source": "solr",
        "request_id": null
    },
    "filters": {
        "title": [{"display","value","link"},...]
    },
    "num_items": "-1",
    "offset": "0"
}
```


### Document Changes
Doc changes included in pd-7092 

#### Associated PR
[backend PR](https://github.com/DirectEmployers/microsites-search-api/pull/151)
added service for single filter endpoint

prevent undefined from appearing in url

Merge branch 'quality-control' of github.com:DirectEmployers/microsite-js into pd-7175-filter-endpoint-service