[PD-7183] - add military base example (#142)
## What does this PR do?
* Adds a military base example using hard coded data to show RR.

### Notes To Reviewer

http://localhost:8080/examples/military-base-map

I just picked a couple of bases from each branch and hard coded them in. I also added miltary bases filter to the config for the chosen bases but made the raw queries target `state_short_exact` instead of `location_exact` because some of the bases show up up out of the exact city locations so this is not accurate but this "fakes" the filters and allows them to show up in the facets.
add military base example

fix example

build